move models stuff to directory
